[REF] web: convert useVirtual tests to HOOT
This commit converts the useVirtual QUnit tests to the HOOT suite.

part of task-3705027
